ML/verification/requirements.txt
Flask
gunicorn
tensorFlow
numpy
Pillow
dlib
scikit-learn
Flask-HTTPAuth
"cloud-sql-python-connector[pymysql]"
google-cloud-sql-connector
sqlalchemy
google-api-python-client
google-cloud-storage
